effective-software-testing/wbk--hello-pytest-w-appium/requirements.txt
appium-python-client
pytest
chromedriver